+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fixes
 
+* `ModuleDirectives` handles even weirder places to hide your aliases (anon functions, in this case)
 * `Pipes` tries even harder to keep single-pipe rewrites of invocations on one line
 
 ## v0.3.1